+++ src/shared/views/Home.tsx
@@ -1,98 +1,117 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { Drawer, notification, Modal, Button, Empty } from 'antd';
-import { Project, PaginatedList, PaginationSettings } from '@bbp/nexus-sdk-legacy';
-import { CreateProjectPayload } from '@bbp/nexus-sdk-legacy/lib/Project/types';
-import { RootState } from '../store/reducers';
-import {
-  modifyProject,
-  createProject,
-  deprecateProject,
-  makeProjectPublic,
-} from '../store/actions/project';
-import ProjectList from '../components/Projects/ProjectList';
-import Skeleton from '../components/Skeleton';
 import { push } from 'connected-react-router';
-import ProjectForm from '../components/Projects/ProjectForm';
-import { fetchOrg } from '../store/actions/nexus/activeOrg';
+import {
+  Organization,
+  PaginatedList,
+  PaginationSettings,
+} from '@bbp/nexus-sdk-legacy';
+import { RootState } from '../store/reducers';
+import { createOrg, modifyOrg, deprecateOrg } from '../store/actions/orgs';
+import OrgList from '../components/Orgs/OrgList';
+import { fetchOrgs } from '../store/actions/nexus/orgs';
+import Skeleton from '../components/Skeleton';
+import { Button, Modal, Drawer, notification, Empty } from 'antd';
+import OrgForm from '../components/Orgs/OrgForm';
+import { CreateOrgPayload } from '@bbp/nexus-sdk-legacy/lib/Organization/types';
+import { Link } from 'react-router-dom';
 
-interface HomeProps {
-  activeOrg: { label: string; description?: string };
-  paginatedProjects: PaginatedList<Project>;
+interface LandingProps {
+  paginatedOrgs?: PaginatedList<Organization>;
   displayPerPage: number;
   busy: boolean;
-  match: any;
-  fetchOrgData(orgLabel: string, paginationSettings?: PaginationSettings): void;
-  createProject(
+  error?: { message: string; name: string };
+  goTo(orgLabel: string): void;
+  fetchOrgs(paginationSettings?: PaginationSettings): any;
+  createOrg: (
     orgLabel: string,
-    projectLabel: string,
-    payload: CreateProjectPayload
-  ): Promise<Project>;
-  modifyProject(
+    orgPayload: CreateOrgPayload
+  ) => Promise<Organization>;
+  modifyOrg: (
     orgLabel: string,
-    projectLabel: string,
     rev: number,
-    payload: CreateProjectPayload
-  ): Promise<Project>;
-  deprecateProject(
-    orgLabel: string,
-    projectLabel: string,
-    rev: number
-  ): Promise<void>;
-  makeProjectPublic(orgLabel: string, projectLabel: string): Promise<void>;
-  goTo(o: string, p: string): void;
+    orgPayload: CreateOrgPayload
+  ) => Promise<Organization>;
+  deprecateOrg: (orgLabel: string, rev: number) => Promise<void>;
 }
 
-const Home: React.FunctionComponent<HomeProps> = ({
+const Landing: React.FunctionComponent<LandingProps> = ({
+  paginatedOrgs = { total: 0, index: 0, results: [] },
   busy,
-  paginatedProjects,
-  match,
-  activeOrg,
-  fetchOrgData,
+  error,
   goTo,
-  createProject,
-  modifyProject,
-  deprecateProject,
-  makeProjectPublic,
+  fetchOrgs,
+  createOrg,
+  modifyOrg,
+  deprecateOrg,
   displayPerPage,
 }) => {
   const [formBusy, setFormBusy] = React.useState<boolean>(false);
   const [modalVisible, setModalVisible] = React.useState<boolean>(false);
-  const [selectedProject, setSelectedProject] = React.useState<
-    Project | undefined
+  const [selectedOrg, setSelectedOrg] = React.useState<
+    Organization | undefined
   >(undefined);
   React.useEffect(() => {
-    if (
-      activeOrg.label !== match.params.org ||
-      (paginatedProjects.results.length === 0 && !busy)
-    ) {
-      fetchOrgData(match.params.org);
-    }
-  }, [match.path]);
+    paginatedOrgs.results.length === 0 &&
+      fetchOrgs({
+        size: displayPerPage,
+        from: paginatedOrgs.index,
+      });
+  }, []);
 
-  const saveAndCreate = (newProject: Project) => {
+  const saveAndCreate = (newOrg: Organization) => {
     setFormBusy(true);
-    createProject(activeOrg.label, newProject.label, {
-      base: newProject.base || undefined,
-      vocab: newProject.vocab || undefined,
-      description: newProject.description || '',
-      apiMappings:
-        newProject.apiMappings.length === 0
-          ? undefined
-          : newProject.apiMappings,
+    createOrg(newOrg.label, { description: newOrg.description || '' })
+      .then(
+        () => {
+          notification.success({
+            message: 'Organization created',
+            duration: 5,
+          });
+          setFormBusy(false);
+          goTo(newOrg.label);
+        },
+        (action: { type: string; error: Error }) => {
+          notification.warning({
+            message: 'Organization NOT created',
+            description: action.error.message,
+            duration: 10,
+          });
+          setFormBusy(false);
+        }
+      )
+      .catch((error: Error) => {
+        notification.error({
+          message: 'An unknown error occurred',
+          description: error.message,
+          duration: 0,
+        });
+      });
+  };
+
+  const saveAndModify = (selectedOrg: Organization, newOrg: Organization) => {
+    setFormBusy(true);
+    modifyOrg(newOrg.label, selectedOrg.rev, {
+      description: newOrg.description,
     })
       .then(
         () => {
           notification.success({
-            message: 'Project created',
+            message: 'Organization saved',
             duration: 2,
           });
           setFormBusy(false);
-          goTo(activeOrg.label, newProject.label);
+          setModalVisible(false);
+          setSelectedOrg(undefined);
+
+          fetchOrgs({
+            size: displayPerPage,
+            from: paginatedOrgs.index,
+          });
         },
         (action: { type: string; error: Error }) => {
           notification.warning({
-            message: 'Project NOT created',
+            message: 'Organization NOT saved',
             description: action.error.message,
             duration: 2,
           });
@@ -108,100 +127,28 @@ const Home: React.FunctionComponent<HomeProps> = ({
       });
   };
 
-  const saveAndModify = (selectedProject: Project, newProject: Project) => {
+  const saveAndDeprecate = (selectedOrg: Organization) => {
     setFormBusy(true);
-    modifyProject(activeOrg.label, newProject.label, selectedProject.rev, {
-      base: newProject.base,
-      vocab: newProject.vocab,
-      description: newProject.description,
-      apiMappings: newProject.apiMappings || [],
-    })
+
+    deprecateOrg(selectedOrg.label, selectedOrg.rev)
       .then(
         () => {
           notification.success({
-            message: 'Project saved',
+            message: 'Organization deprecated',
             duration: 2,
           });
           setFormBusy(false);
           setModalVisible(false);
-          setSelectedProject(undefined);
+          setSelectedOrg(undefined);
 
-          fetchOrgData(match.params.org);
+          fetchOrgs({
+            size: displayPerPage,
+            from: paginatedOrgs.index,
+          });
         },
         (action: { type: string; error: Error }) => {
           notification.warning({
-            message: 'Project NOT saved',
-            description: action.error.message,
-            duration: 2,
-          });
-          setFormBusy(false);
-        }
-      )
-      .catch((error: Error) => {
-        notification.error({
-          message: 'An unknown error occurred',
-          description: error.message,
-          duration: 0,
-        });
-      });
-  };
-
-  const saveAndDeprecate = (selectedProject: Project) => {
-    setFormBusy(true);
-
-    deprecateProject(
-      selectedProject.orgLabel,
-      selectedProject.label,
-      selectedProject.rev
-    )
-      .then(
-        () => {
-          notification.success({
-            message: 'Project successfully deprecated',
-            duration: 2,
-          });
-          setFormBusy(false);
-          setModalVisible(false);
-          setSelectedProject(undefined);
-
-          fetchOrgData(match.params.org);
-        },
-        (action: { type: string; error: Error }) => {
-          notification.warning({
-            message: 'Project NOT deprecated',
-            description: action.error.message,
-            duration: 2,
-          });
-          setFormBusy(false);
-        }
-      )
-      .catch((error: Error) => {
-        notification.error({
-          message: 'An unknown error occurred',
-          description: error.message,
-          duration: 0,
-        });
-      });
-  };
-
-  const makePublic = (selectedProject: Project) => {
-    setFormBusy(true);
-    makeProjectPublic(selectedProject.orgLabel, selectedProject.label)
-      .then(
-        () => {
-          notification.success({
-            message: 'Project is now publicly accessible',
-            duration: 2,
-          });
-          setFormBusy(false);
-          setModalVisible(false);
-          setSelectedProject(undefined);
-
-          fetchOrgData(match.params.org);
-        },
-        (action: { type: string; error: Error }) => {
-          notification.warning({
-            message: 'Project NOT made public',
+            message: 'Organization NOT deprecated',
             description: action.error.message,
             duration: 2,
           });
@@ -234,79 +181,91 @@ const Home: React.FunctionComponent<HomeProps> = ({
     );
   }
 
+  if (error) {
+    if (error.message === 'Error: Forbidden') {
+      return (
+        <Empty
+          style={{ marginTop: '22vh' }}
+          description={
+            <span>
+              You need to <Link to="/login">login</Link> in order to list
+              Organizations
+            </span>
+          }
+        />
+      );
+    }
+    return (
+      <Empty
+        style={{ marginTop: '22vh' }}
+        description="There was a problem while loading Organizations"
+      />
+    );
+  }
+
   return (
-    <div className="projects-view">
-      <h1 style={{ marginBottom: 0, marginRight: 8 }}>{activeOrg.label}</h1>
-      {activeOrg.description && <p>{activeOrg.description}</p>}
+    <div className="orgs-view">
       <div style={{ display: 'flex', alignItems: 'center', marginBottom: 20 }}>
-        <h2 style={{ marginBottom: 0, marginRight: 8 }}>Projects</h2>
+        <h1 style={{ marginBottom: 0, marginRight: 8 }}>Organizations</h1>
         <Button
           type="primary"
           onClick={() => setModalVisible(true)}
           icon="plus-square"
         >
-          Create Project
+          Create Organization
         </Button>
       </div>
-      {paginatedProjects.total === 0 ? (
-        <Empty description="No projects" />
+      {paginatedOrgs.total === 0 ? (
+        <Empty
+          style={{ marginTop: '22vh' }}
+          description="No Organizations found..."
+        />
       ) : (
-        <ProjectList
-          projects={paginatedProjects.results}
-          onProjectClick={(projectLabel: string) =>
-            goTo(activeOrg.label, projectLabel)
-          }
-          onProjectEdit={(projectLabel: string) =>
-            setSelectedProject(
-              paginatedProjects.results.filter(p => p.label === projectLabel)[0]
-            )
-          }
+        <OrgList
+          orgs={paginatedOrgs.results}
+          onOrgClick={goTo}
+          busy={busy}
           paginationSettings={{
-            total: paginatedProjects.total,
-            from: paginatedProjects.index,
+            total: paginatedOrgs.total,
+            from: paginatedOrgs.index,
             pageSize: displayPerPage,
           }}
+          onOrgEdit={(orgLabel: string) =>
+            setSelectedOrg(
+              paginatedOrgs.results.filter(o => o.label === orgLabel)[0]
+            )
+          }
           onPaginationChange={pageNumber =>
-            fetchOrgData(match.params.org, {
-              from: displayPerPage * pageNumber - displayPerPage,
+            fetchOrgs({
+              from: pageNumber * displayPerPage - displayPerPage,
               size: displayPerPage,
             })
           }
-          busy={busy}
         />
       )}
       <Modal
-        title="New Project"
+        title="New Organization"
         visible={modalVisible}
         onCancel={() => setModalVisible(false)}
         confirmLoading={formBusy}
         footer={null}
-        width={600}
       >
-        <ProjectForm
-          onSubmit={(p: Project) => saveAndCreate(p)}
+        <OrgForm
+          onSubmit={(o: Organization) => saveAndCreate(o)}
           busy={formBusy}
         />
       </Modal>
       <Drawer
         width={640}
-        visible={!!(selectedProject && selectedProject.label)}
-        onClose={() => setSelectedProject(undefined)}
-        title={`Project: ${selectedProject && selectedProject.label}`}
+        visible={!!(selectedOrg && selectedOrg.label)}
+        onClose={() => setSelectedOrg(undefined)}
+        title={selectedOrg && selectedOrg.label}
       >
-        {selectedProject && (
-          <ProjectForm
-            project={{
-              label: selectedProject.label,
-              rev: selectedProject.rev,
-              description: selectedProject.description || '',
-              base: selectedProject.base,
-              vocab: selectedProject.vocab,
-              apiMappings: selectedProject.apiMappings,
-            }}
-            onSubmit={(p: Project) => saveAndModify(selectedProject, p)}
-            onDeprecate={() => saveAndDeprecate(selectedProject)}
-            onMakePublic={() => makePublic(selectedProject)}
+        {selectedOrg && (
+          <OrgForm
+            org={selectedOrg}
+            onSubmit={(o: Organization) => saveAndModify(selectedOrg, o)}
+            onDeprecate={() => saveAndDeprecate(selectedOrg)}
             busy={formBusy}
             mode="edit"
           />
@@ -317,44 +276,25 @@ const Home: React.FunctionComponent<HomeProps> = ({
 };
 
 const mapStateToProps = (state: RootState) => ({
-  displayPerPage: state.uiSettings.pageSizes.projectsListPageSize,
-  activeOrg: (state.nexus &&
-    state.nexus.activeOrg &&
-    state.nexus.activeOrg.data &&
-    state.nexus.activeOrg.data.org) || { label: '' },
-  paginatedProjects: (state.nexus &&
-    state.nexus.activeOrg &&
-    state.nexus.activeOrg.data &&
-    state.nexus.activeOrg.data.projects) || { results: [], total: 0, index: 0 },
-  busy:
-    (state.nexus &&
-      state.nexus.activeOrg &&
-      state.nexus.activeOrg.isFetching) ||
-    false,
+  displayPerPage: state.uiSettings.pageSizes.orgsListPageSize,
+  paginatedOrgs: (state.nexus && state.nexus.orgs.data) || undefined,
+  busy: (state.nexus && state.nexus.orgs.isFetching) || false,
+  error: (state.nexus && state.nexus.orgs.error) || undefined,
 });
 
 const mapDispatchToProps = (dispatch: any) => ({
-  fetchOrgData: (orgLabel: string, paginationSettings?: PaginationSettings) =>
-    dispatch(fetchOrg(orgLabel, paginationSettings)),
-  goTo: (org: string, project: string) => dispatch(push(`/${org}/${project}`)),
-  createProject: (
-    orgLabel: string,
-    projectLabel: string,
-    payload: CreateProjectPayload
-  ) => dispatch(createProject(orgLabel, projectLabel, payload)),
-  modifyProject: (
-    orgLabel: string,
-    projectLabel: string,
-    rev: number,
-    payload: CreateProjectPayload
-  ) => dispatch(modifyProject(orgLabel, projectLabel, rev, payload)),
-  deprecateProject: (orgLabel: string, projectLabel: string, rev: number) =>
-    dispatch(deprecateProject(orgLabel, projectLabel, rev)),
-  makeProjectPublic: (orgLabel: string, projectLabel: string) =>
-    dispatch(makeProjectPublic(orgLabel, projectLabel)),
+  goTo: (org: string) => dispatch(push(`/${org}`)),
+  fetchOrgs: (paginationSettings?: PaginationSettings) =>
+    dispatch(fetchOrgs(paginationSettings)),
+  createOrg: (orgLabel: string, orgPayload: CreateOrgPayload) =>
+    dispatch(createOrg(orgLabel, orgPayload)),
+  modifyOrg: (orgLabel: string, rev: number, orgPayload: CreateOrgPayload) =>
+    dispatch(modifyOrg(orgLabel, rev, orgPayload)),
+  deprecateOrg: (orgLabel: string, rev: number) =>
+    dispatch(deprecateOrg(orgLabel, rev)),
 });
 
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(Home);
+)(Landing);
